Added labels, still need a bit of work

+++ Sprite.cpp
@@ -64,7 +64,7 @@ int Sprite::getHeight() {
 }
 
 SDL_Rect Sprite::getRect() {
-	SDL_Rect rect = { x, y, width, height };
+	SDL_Rect rect = { (int)x, (int)y, width, height };
 	return rect;
 }
